% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baidumap-package.R
\docType{package}
\name{baidumap-package}
\alias{baidumap}
\alias{baidumap-package}
\title{baidumap: A package for spatial visualization with Baidu}
\description{
Just like ggmap, but get map from baidu instead of google or openstreet
}
\keyword{internal}
