% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoconv.R
\name{geoconv}
\alias{geoconv}
\title{Convert geocode
Take in geocode from the other source to baidu's geocode}
\usage{
geoconv(geocode, from = 3, to = 5)
}
\arguments{
\item{from}{takes intergers from 1 to 8. See more in details.}

\item{to}{5 or 6. See more in details.}

\item{geocde}{geocode from the other source}
}
\description{
Convert geocode
Take in geocode from the other source to baidu's geocode
}
\details{
Go to http://developer.baidu.com/map/index.php?title=webapi/guide/changeposition to see what the intergers mean.
}

