% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coord.R
\name{get_coord}
\alias{get_coord}
\title{Get coordiante from address
Take in address and return the coordinate}
\usage{
get_coord(address, city = NULL, map_ak = "", limit = 600, .parallel = NULL)
}
\arguments{
\item{address}{address}

\item{city}{the city of the address, optional}

\item{limit}{integer value.If the length of address exceeded limit, function will run in parallel}

\item{output}{should be "json" or "xml", the type of the result}

\item{formatted}{logical value, return the coordinates or the original results}
}
\value{
A vector contains the  corresponding coordiante. If "formatted=TRUE", 
return the numeric coordinates, otherwise return json or xml type result, 
depents on the argument "output". If the length of address is larger than 1, 
the result is a matrix.
}
\description{
Get coordiante from address
Take in address and return the coordinate
}
\references{
1. http://lbsyun.baidu.com/index.php?title=webapi/guide/webservice-geocoding
}
