% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location.R
\name{get_location}
\alias{get_location}
\title{Get location from coordinate
Take in coordiantes and return the location}
\usage{
get_location(location, limit = 600, map_ak = "")
}
\arguments{
\item{location}{longtitude and latitude}

\item{limit}{integer value.If the number of row exceeded limit, function will run in parallel}
}
\value{
the corresponding locations
}
\description{
Get location from coordinate
Take in coordiantes and return the location
}
\examples{
\dontrun{
## get one location
location_one <- get_location(c(118.12845, 24.57742))

## vectorization
loc <- matrix(c(117.93780, 24.55730, 117.93291, 24.57745, 117.23530, 24.64210, 117.05890, 24.74860), byrow = T, ncol = 2)
get_location(loc)
}
}
\references{
1. http://lbs.baidu.com/index.php?title=webapi/guide/webservice-geocoding-abroad
}
