% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_place.R
\name{get_place}
\alias{get_place}
\title{Get place from query
Take in query and city, return the informations}
\usage{
get_place(
  place = NULL,
  city = "",
  page_size = 20,
  pages = Inf,
  scope = 1,
  verbose = TRUE,
  map_ak = ""
)
}
\arguments{
\item{place}{the place you want to search}

\item{city}{define the city}
}
\value{
a data frame contains name, longtitude, latitude and address, as well as teleplhone number, if exist.
}
\description{
Get place from query
Take in query and city, return the informations
}
\examples{
\dontrun{
## colleges in beijing
bj_college = get_place('澶у', '鍖椾含')
## Mcdonald's in shanghai
sh_mcdonald = get_place('楹﹀綋鍔戼㸳', '涓婃捣')
}
}
