% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBaiduMap}
\alias{getBaiduMap}
\title{Get map from Baidu API}
\usage{
getBaiduMap(lon = 116.354431, lat = 39.942333, width = 400,
  height = 400, zoom = 10, scale = 2, color = "color",
  messaging = TRUE)
}
\arguments{
\item{lon}{longtitude of the center of the map}

\item{lat}{latitude of the center of the map}

\item{width}{width of the map}

\item{height}{height of the map}

\item{zoom}{map zoom, an integer from 3 (continent) to 21 (building), default value 10 (city)}

\item{scale}{multiplicative factor for the number of pixels returned. possible values are 2 or anything else.}

\item{color}{"color" or "bw", color or black-and-white}

\item{messaging}{logical. whether to print the download messages.}
}
\value{
A ggmap object. a map image as a 2d-array of colors as hexadecimal strings representing pixel fill values.
}
\description{
Take in coordiantes and return the location
}
\examples{
\dontrun{
library(ggmap)
## default map: Beijing
p <- getBaiduMap()
ggmap(p)

## black-and-white
p <- getBaiduMap(color='bw')
ggmap(p)

## do not print messages
p <- getBaiduMap(messaging = F)
}
}

