% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLocation}
\alias{getLocation}
\title{Get location from coordinate
Take in coordiantes and return the location}
\usage{
getLocation(lon, lat, output = "json", formatted = F, pois = 0)
}
\arguments{
\item{lon}{longtitude}

\item{lat}{latitude}

\item{output}{should be "json" or "xml", the type of the result}

\item{formatted}{logical. whether to return a nice result}

\item{pois}{whether to return the PIO around the location}
}
\value{
the corresponding locations
}
\description{
Get location from coordinate
Take in coordiantes and return the location
}
\examples{
\dontrun{
## get one location
location_one = getLocation(118.12845, 24.57742)

## vectorization
lon = matrix(c(117.93780, 24.55730, 117.93291, 24.57745, 117.23530, 24.64210, 117.05890, 24.74860), byrow=T, ncol=2)
### json
location_json = getLocation(lon[, 1], lon[, 2], output='json')
### get district
library(rjson)
getDistrict = function(x_json){
    x_list = fromJSON(x_json)
    x_list$result$addressComponent$district
}
location_district = sapply(location_json, getDistrict)

### xml
location_xml = getLocation(lon[, 1], lon[, 2], output='xml')

## formatted
location = getLocation(lon[, 1], lon[, 2], formatted = T)
}
}

