% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoute.R
\name{getRoute}
\alias{getRoute}
\title{Get route from query
Take in a original location and destination, return the direction}
\usage{
getRoute(...)
}
\arguments{
\item{origin}{the original location}

\item{destination}{the destination}

\item{mode}{'driving'(default), 'walking', or 'transit'}

\item{region}{the city of of original location and destination. 
If original and destination is not in the same city, 
set `origin_region` and `destination_region` seperately.}

\item{origin_region}{the city of original location. If not set, use `region` instead.}

\item{destination_region}{the city of destination. If not set, use `region` instead.}

\item{tactics}{10(no expressway), 11(default, shortest time), 12(shortest path).}

\item{coord_type}{'bd09ll'(default), 'gcj02'(which Google map and Soso map are using), 'wgs84' for GPS devices.}
}
\value{
a data frame contains longtitude and latitude of the route.
}
\description{
Get route from query
Take in a original location and destination, return the direction
}
\examples{
\dontrun{
bjMap = getBaiduMap('北京', color='bw')
df = getRoute('首都国际机场', '北京南苑机场', region = '北京')
ggmap(bjMap) + geom_path(data = df, aes(lon, lat), alpha = 0.5, col = 'red')
}

}

