% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlace.R
\name{url_character}
\alias{url_character}
\title{Transform the query character to raw character
Take in query and city, return the informations}
\usage{
url_character(x)
}
\arguments{
\item{a}{character}
}
\value{
raw character with %. It's used in getPlace.
}
\description{
Transform the query character to raw character
Take in query and city, return the informations
}
\examples{

\dontrun{
url_character('北京')
# "\%e5\%8c\%97\%e4\%ba\%ac"
}
}

