% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaiduMap.R
\name{getBaiduMap}
\alias{getBaiduMap}
\title{Get map from Baidu API}
\usage{
getBaiduMap(location, width = 400, height = 400, zoom = 10, scale = 2,
  color = "color", messaging = TRUE)
}
\arguments{
\item{location}{a vector a or matrix contains longtitude and latitude of the center of the map, or a character refers to the address}

\item{width}{width of the map}

\item{height}{height of the map}

\item{zoom}{map zoom, an integer from 3 (continent) to 21 (building), default value 10 (city)}

\item{scale}{multiplicative factor for the number of pixels returned. possible values are 2 or anything else.}

\item{color}{"color" or "bw", color or black-and-white}

\item{messaging}{logical. whether to print the download messages.}
}
\value{
A ggmap object. a map image as a 2d-array of colors as hexadecimal strings representing pixel fill values.
}
\description{
Take in coordiantes and return the location
}
\examples{

\dontrun{
library(ggmap)  
## Beijing
p <- getBaiduMap(c(116.39565, 39.92999))
ggmap(p)

p <- getBaiduMap('beijing') # the same
ggmap(p)

## black-and-white
p <- getBaiduMap(color='bw')
ggmap(p)

## do not print messages
p <- getBaiduMap(messaging = F)
}
}

