% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordinate.R
\name{getCoordinate}
\alias{getCoordinate}
\title{Get coordiante from address
Take in address and return the coordinate}
\usage{
getCoordinate(address, city = NULL, output = "json", formatted = F,
  limit = 600)
}
\arguments{
\item{address}{address}

\item{city}{the city of the address, optional}

\item{output}{should be "json" or "xml", the type of the result}

\item{formatted}{logical value, return the coordinates or the original results}

\item{limit}{integer value.If the length of address exceeded limit, function will run in parallel}
}
\value{
A vector contains the  corresponding coordiante. If "formatted=TRUE", return the numeric coordinates, otherwise return json or xml type result, depents on the argument "output". If the length of address is larger than 1, the result is a matrix.
}
\description{
Get coordiante from address
Take in address and return the coordinate
}
\examples{
\dontrun{ 
## json output
getCoordinate('北京大学')

## xml output
getCoordinate('北京大学', output='xml')

## formatted
getCoordinate('北京大学', formatted = T)

## vectorization, return a matrix
getCoordinate(c('北京大学', '清华大学'), formatted = T)
}
}

