% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocation.R
\name{getLocation}
\alias{getLocation}
\title{Get location from coordinate
Take in coordiantes and return the location}
\usage{
getLocation(location, output = "json", formatted = F, pois = 0,
  limit = 600)
}
\arguments{
\item{location}{longtitude and latitude}

\item{output}{should be "json" or "xml", the type of the result}

\item{formatted}{logical. whether to return a nice result}

\item{pois}{whether to return the POI around the location}

\item{limit}{integer value.If the number of row exceeded limit, function will run in parallel}
}
\value{
the corresponding locations
}
\description{
Get location from coordinate
Take in coordiantes and return the location
}
\examples{

\dontrun{  
## get one location 
location_one = getLocation(c(118.12845, 24.57742))

## vectorization
loc = matrix(c(117.93780, 24.55730, 117.93291, 24.57745, 117.23530, 24.64210, 117.05890, 24.74860), byrow=T, ncol=2)
### json 
location_json = getLocation(loc, output='json')
### get district
library(rjson)
getDistrict = function(x_json){
    x_list = fromJSON(x_json)
    x_list$result$addressComponent$district
}
location_district = sapply(location_json, getDistrict)

### xml
location_xml = getLocation(loc, output='xml')

## formatted
location = getLocation(loc, formatted = T) 
}

}

